Simplified logging
All packages use `cerbero.Stdout` and `cerbero.Stderr`. This simplify the configuration, with one place to configure all the loggers.

The idea behind this change is that while backtesting we don't need any logs and we want to go fast. While running, we want all the infos to understand if the system is working properly